+++ Entities/TransactionHistory.cs
@@ -8,7 +8,7 @@ namespace BankingApplication.Entities
         public ObjectId Id { get; set; }
         public DateTime DateTime { get; set; }
         public TransactionType Type { get; set; }
-        public decimal Amount { get; set; }
+        public int Amount { get; set; }
         public string? ReceiverAccountNumber { get; set; }
         public string? SenderAccountNumber { get; set; }
         public ObjectId AccountId { get; set; }
